// shiku-world-medium/client/communication/api/bindings/GuestInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface GuestInput { jump: boolean, up: boolean, down: boolean, left: boolean, right: boolean, start: boolean, action_1: boolean, action_2: boolean, x_axis: number, y_axis: number, }